+++ yandex-mapkit-kmp-moko-compose/src/commonMain/kotlin/ru/sulgik/mapkit/moko/compose/moko.kt
@@ -1,7 +1,0 @@
-package ru.sulgik.mapkit.moko.compose
-
-import androidx.compose.runtime.Composable
-import ru.sulgik.mapkit.moko.MOKOImageLoader
-
-@Composable
-public expect fun rememberMOKOImageLoader(): MOKOImageLoader