+++ yandex-mapkit-kmp/src/iosMain/kotlin/ru/sulgik/mapkit/map/Map.ios.kt
@@ -198,6 +198,9 @@ public actual class Map internal constructor(private val nativeMap: NativeMap) {
 
     /**
      * Adds camera listeners.
+     *
+     * The class does not retain the object in the 'cameraListener' parameter.
+     * It is your responsibility to maintain a strong reference to the target object while it is attached to a class.
      */
     public actual fun addCameraListener(cameraListener: WeakRef<CameraListener>) {
         nativeMap.addCameraListenerWithCameraListener(cameraListener.toNative() ?: return)
@@ -227,6 +230,9 @@ public actual class Map internal constructor(private val nativeMap: NativeMap) {
 
     /**
      * Adds input listeners.
+     *
+     * The class does not retain the object in the 'inputListener' parameter.
+     * It is your responsibility to maintain a strong reference to the target object while it is attached to a class.
      */
     public actual fun addInputListener(inputListener: WeakRef<InputListener>) {
         nativeMap.addInputListenerWithInputListener(inputListener.toNative() ?: return)
@@ -300,7 +306,7 @@ public actual class Map internal constructor(private val nativeMap: NativeMap) {
     /**
      * Selects one of predefined map style modes optimized for particular use case(transit, driving, etc).
      *
-     * Resets json styles set with setMapStyle. MapMode.Map by deafult.
+     * Resets json styles set with setMapStyle. MapMode.Map by default.
      */
     public actual var mode: MapMode
         get() = nativeMap.mode.toCommon()