+++ yandex-mapkit-kmp/src/iosMain/kotlin/ru/sulgik/mapkit/runtime/Error.ios.kt
@@ -1,12 +1,5 @@
 package ru.sulgik.mapkit.runtime
 
-import ru.sulgik.mapkit.runtime.network.BadRequestError
-import ru.sulgik.mapkit.runtime.network.ForbiddenError
-import ru.sulgik.mapkit.runtime.network.NetworkError
-import ru.sulgik.mapkit.runtime.network.NotFoundError
-import ru.sulgik.mapkit.runtime.network.RemoteError
-import ru.sulgik.mapkit.runtime.network.RequestEntityTooLargeError
-import ru.sulgik.mapkit.runtime.network.UnauthorizedError
 import ru.sulgik.mapkit.runtime.network.toCommon
 import YandexMapKit.YRTBadRequestError as NativeBadRequestError
 import YandexMapKit.YRTDiskCorruptError as NativeDiskCorruptError